+++ yapw/util.py
@@ -1,5 +1,7 @@
 import json
 
+import pika
+
 try:
     import orjson
 
@@ -15,3 +17,19 @@ def json_dumps(data):
     if jsonlib == json:
         return json.dumps(data, separators=(",", ":")).encode()
     return orjson.dumps(data)
+
+
+def basic_publish_kwargs(state, message, routing_key):
+    """
+    Prepares keyword arguments for ``basic_publish``.
+
+    :param state: an object with ``format_routing_key``, ``exchange`` and ``delivery_mode`` attributes
+    :param message: a JSON-serializable message
+    :param str routing_key: the routing key
+    """
+    formatted = state.format_routing_key(routing_key)
+
+    body = json_dumps(message)
+    properties = pika.BasicProperties(delivery_mode=state.delivery_mode, content_type="application/json")
+
+    return {"exchange": state.exchange, "routing_key": formatted, "body": body, "properties": properties}